Se arreglan bug de los filtros

+++ src/main/java/com/soin/sgrm/service/RFCService.java
@@ -3,6 +3,7 @@ package com.soin.sgrm.service;
 import java.sql.SQLException;
 
 import com.soin.sgrm.model.RFC;
+import com.soin.sgrm.model.RFCList;
 import com.soin.sgrm.response.JsonSheet;
 
 public interface RFCService extends BaseService<Long, RFC>{
@@ -10,11 +11,11 @@ public interface RFCService extends BaseService<Long, RFC>{
 	
 	String generateRFCNumber(String codeProject);
 	JsonSheet<RFC> findAll1(Integer sEcho, Integer iDisplayStart, Integer iDisplayLength, String sSearch, Long statusId,
-			String dateRange, Long priorityId,Long impactId);
+			String dateRange, int priorityId,int impactId);
 
 	Integer countByType(String name, String type, int query, Object[] ids);
 
 	public JsonSheet<RFC> findAll2(String name, Integer sEcho, Integer iDisplayStart, Integer iDisplayLength,
-			String sSearch, Long statusId, String dateRange, Long priorityId, Long impactId);
+			String sSearch, Long statusId, String dateRange, int priorityId, int impactId);
 
 }
